+++ R/in_case_fct.R
@@ -15,8 +15,16 @@
 #'
 #' @return A factor vector of length 1 or n, matching the length of the logical
 #'   input or output vectors.
-#'   Levels are determined by the order of inputs to `...`.
+#'   Levels are determined by the order of inputs to `...` and `.default`.
 #'   Inconsistent lengths will generate an error.
+#'
+#'   The position of the `.default` argument is taken into account when setting
+#'   factor levels in `*_case_fct()` functions.
+#'   For example, if the `.default` argument is given before any case
+#'   statements, the default value will be the first level of the factor;
+#'   if the `.default` argument is positioned in between two case statements,
+#'   the default value will be ordered in between the value of the two
+#'   case statements.
 #'
 #' @seealso [in_case()], [switch_case()], [grep_case()], [fn_case()], and
 #'   [fn_case_fct()] on which these functions are based.
@@ -37,7 +45,11 @@ in_case_fct <- function(
   .default <- coalesce_deprecated(.default, default)
   .ordered <- coalesce_deprecated(.ordered, ordered)
 
+  call <- rlang::current_call()
   dots <- compact_list(...)
+  dots_idx <- which(as.list(call) %in% dots)
+  .default_idx <- which(names(call) %in% c("default", ".default"))
+
   inputs <- in_case_setup(dots, .preserve = .preserve, fn = "in_case_fct")
 
   replace(
@@ -48,7 +60,9 @@ in_case_fct <- function(
     factor      = TRUE,
     .ordered    = .ordered,
     default_env = rlang::caller_env(),
-    current_env = rlang::current_env()
+    current_env = rlang::current_env(),
+    dots_idx = dots_idx,
+    .default_idx = .default_idx
   )
 }
 
@@ -69,14 +83,10 @@ switch_case_fct <- function(
   .default <- coalesce_deprecated(.default, default)
   .ordered <- coalesce_deprecated(.ordered, ordered)
 
-  fn_case_fct(
-    x  = x,
-    fn = `%in%`,
-    ...,
-    .preserve = .preserve,
-    .default  = .default,
-    .ordered  = .ordered
-  )
+  args <- as.list(rlang::current_call())[-1]
+  args <- dot_arg_names(args)
+
+  eval.parent(rlang::call2("fn_case_fct", fn = `%in%`, !!!args))
 }
 
 #' @rdname in_case_fct
@@ -96,14 +106,10 @@ grep_case_fct <- function(
   .default <- coalesce_deprecated(.default, default)
   .ordered <- coalesce_deprecated(.ordered, ordered)
 
-  fn_case_fct(
-    x  = x,
-    fn = grepl_any,
-    ...,
-    .preserve = .preserve,
-    .default  = .default,
-    .ordered  = .ordered
-  )
+  args <- as.list(rlang::current_call())[-1]
+  args <- dot_arg_names(args)
+
+  eval.parent(rlang::call2("fn_case_fct", fn = grepl_any, !!!args))
 }
 
 #' @rdname in_case_fct
@@ -124,15 +130,21 @@ fn_case_fct <- function(
   .default <- coalesce_deprecated(.default, default)
   .ordered <- coalesce_deprecated(.ordered, ordered)
 
+  call <- rlang::current_call()
   dots <- compact_list(...)
+  dots_idx <- which(as.list(call) %in% dots)
+  .default_idx <- which(names(call) %in% c("default", ".default"))
+
   inputs <- fn_case_setup(dots)
 
   replace(
     inputs$fs, x, .default, .preserve, fn, inputs$args,
-    factor      = TRUE,
-    .ordered    = .ordered,
+    factor = TRUE,
+    .ordered = .ordered,
     default_env = rlang::caller_env(),
-    current_env = rlang::current_env()
+    current_env = rlang::current_env(),
+    dots_idx = dots_idx,
+    .default_idx = .default_idx
   )
 }
 
@@ -154,18 +166,17 @@ fn_switch_case_fct <- function(
   .default <- coalesce_deprecated(.default, default)
   .ordered <- coalesce_deprecated(.ordered, ordered)
 
-  inputs <- fn_switch_case_setup(
-    ...,
-    fn          = fn,
-    default_env = rlang::caller_env(),
-    current_env = rlang::current_env()
+  eval.parent(fn_switch_case_call("switch_case_fct", fn, ...))
+}
+
+dot_arg_names <- function(args) {
+  rlang::names2(args) <- switch_case(
+    rlang::names2(args),
+    "preserve" ~ ".preserve",
+    "default" ~ ".default",
+    "ordered" ~ ".ordered",
+    .preserve = TRUE
   )
 
-  do.call(
-    switch_case_fct,
-    c(
-      list(x = x), inputs$fs, inputs$args,
-      list(.preserve = .preserve, .default = .default, .ordered = .ordered)
-    )
-  )
+  args
 }