+++ R/in_case_list.R
@@ -126,18 +126,5 @@ fn_switch_case_list <- function(
   .preserve <- coalesce_deprecated(.preserve, preserve)
   .default <- coalesce_deprecated(.default, default)
 
-  inputs <- fn_switch_case_setup(
-    ...,
-    fn          = fn,
-    default_env = rlang::caller_env(),
-    current_env = rlang::current_env()
-  )
-
-  do.call(
-    switch_case_list,
-    c(
-      list(x = x), inputs$fs, inputs$args,
-      list(.preserve = .preserve, .default = .default)
-    )
-  )
+  eval.parent(fn_switch_case_call("switch_case_list", fn, ...))
 }